Refactor everything: - fully utilize TCPSock for shared methods
NOTE: RT methods (which addtimers) all need to be PUBLIC
